+++ Project/Views/Model/EmployeeDTO.cs
@@ -9,7 +9,7 @@ namespace Project.Views.Model
 {
     public class EmployeeDTO : UserDTO
     {
-        public double Salary { get; set; }
+        public int Salary { get; set; }
         public TimeInterval AnnualLeave { get; set; }
         public TimeInterval WorkingHours { get; set; }
         public string Email { get; set; }
@@ -17,7 +17,7 @@ namespace Project.Views.Model
         public string Hospital { get; set; }
         public EmployeeDTO() { }
         
-        public EmployeeDTO(long id, AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, double salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password,string hospital) :
+        public EmployeeDTO(long id, AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, int salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password,string hospital) :
               base(id, address, firstName, lastName, jmbg, telephoneNumber, gender, dateOfBirth)
         {
             Salary = salary;
@@ -27,7 +27,7 @@ namespace Project.Views.Model
             Password = password;
             Hospital = hospital;
         }
-        public EmployeeDTO(AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, double salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password,string hospital) :
+        public EmployeeDTO(AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, int salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password,string hospital) :
               base(address, firstName, lastName, jmbg, telephoneNumber, gender, dateOfBirth)
         {
             Salary = salary;
@@ -38,7 +38,7 @@ namespace Project.Views.Model
             Hospital = hospital;
         }
 
-        public EmployeeDTO(long id, AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, double salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password) :
+        public EmployeeDTO(long id, AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, int salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password) :
               base(id, address, firstName, lastName, jmbg, telephoneNumber, gender, dateOfBirth)
         {
             Salary = salary;
@@ -47,7 +47,7 @@ namespace Project.Views.Model
             Email = email;
             Password = password;
         }
-        public EmployeeDTO(AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, double salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password) :
+        public EmployeeDTO(AddressDTO address, string firstName, string lastName, string jmbg, string telephoneNumber, string gender, DateTime dateOfBirth, int salary, TimeInterval annualLeave, TimeInterval workingHours, string email, string password) :
               base(address, firstName, lastName, jmbg, telephoneNumber, gender, dateOfBirth)
         {
             Salary = salary;
